Move types to interfaces directories
