+++ express-back-end/src/api/auth/controllers/auth.controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { UserRequest } from '../types';
+import UserRequest from '../interfaces/UserRequest';
 import {
   EmailTaken,
   Signup,
