+++ express-back-end/src/api/auth/interfaces/User.ts
@@ -1,0 +1,5 @@
+export default interface User {
+  _id: string;
+  email: string;
+  name: string;
+}
