+++ express-back-end/src/api/auth/interfaces/UserDocument.ts
@@ -1,0 +1,7 @@
+import { Document } from 'mongoose';
+
+export default interface UserDocument extends Document {
+  name: string;
+  email: string;
+  password: string;
+}
