+++ express-back-end/src/api/auth/interfaces/UserRequest.ts
@@ -1,0 +1,6 @@
+import { Request } from 'express';
+import User from './User';
+
+export default interface UserRequest extends Request {
+  user?: User;
+}
