--- express-back-end/src/api/auth/types.ts
@@ -1,11 +1,0 @@
-import { Request } from 'express';
-
-export interface User {
-  _id: string;
-  email: string;
-  name: string;
-}
-
-export interface UserRequest extends Request {
-  user?: User;
-}
