+++ express-back-end/src/api/timeline/interfaces/FrequencyData.ts
@@ -1,0 +1,9 @@
+export default interface FrequencyData {
+  word: string;
+  data: {
+    year: number;
+    rank: number;
+    count: number;
+    freq: number;
+  }[];
+}
