+++ express-back-end/src/api/timeline/interfaces/FrequencyDocument.ts
@@ -1,0 +1,9 @@
+import { Document } from 'mongoose';
+
+export default interface FrequencyDocument extends Document {
+  word: string;
+  year: number;
+  rank: number;
+  count: number;
+  freq: number;
+}
