+++ express-back-end/src/api/timeline/interfaces/LatentAssociationData.ts
@@ -1,0 +1,4 @@
+export default interface LatentAssociationData {
+  yearRange: string;
+  association: number;
+}
