+++ express-back-end/src/api/timeline/interfaces/LatentAssociationDocument.ts
@@ -1,0 +1,10 @@
+import { Document } from 'mongoose';
+
+export default interface LatentAssociationDocument extends Document {
+  word: string;
+  vectors: number[];
+  // eslint-disable-next-line camelcase
+  year_from: number;
+  // eslint-disable-next-line camelcase
+  year_to: number;
+}
